refactor(utils): segregate utils into v1, v2, shared.
