+++ src/app.js
@@ -1,15 +1,15 @@
 import express from "express";
 import cors from "cors";
 import cookieParser from "cookie-parser";
-import { errorHandler } from "./utils/errorHandler.js";
+import { errorHandler } from "./utils/shared/errorHandler.js";
 import v1Routes from "./routes/v1/routes.config.js";
 import v2Routes from "./routes/v2/routes.config.js";
 import { STATUS } from "./constants/statusCodes.js";
 import { morganMiddleware } from "./middlewares/shared/morgan.middleware.js";
-import "./utils/instrument.js";
+import "./utils/shared/instrument.js";
 import * as Sentry from "@sentry/node";
-import { limiter } from "./utils/rateLimiter.js";
-import { ApiResponse } from "./utils/apiResponse.js";
+import { limiter } from "./utils/shared/rateLimiter.js";
+import { ApiResponse } from "./utils/shared/apiResponse.js";
 
 const app = express();
 Sentry.setupExpressErrorHandler(app);
