+++ src/controllers/v1/deliveryPartner.controller.js
@@ -1,10 +1,10 @@
-import { asyncHandler } from "../../utils/asyncHandler.js";
-import { ApiResponse } from "../../utils/apiResponse.js";
+import { asyncHandler } from "../../utils/shared/asyncHandler.js";
+import { ApiResponse } from "../../utils/shared/apiResponse.js";
 import bcrypt from "bcrypt";
 import {
   generateAccessToken,
   generateRefreshToken,
-} from "../../utils/tokenGenerator.js";
+} from "../../utils/v1/tokenGenerator.js";
 import {
   createDeliveryPartner,
   deleteDeliveryPartner,
@@ -17,9 +17,9 @@ import {
 } from "../../database/v1/queries/deliveryPartner.query.js";
 import jwt from "jsonwebtoken";
 import fs from "fs";
-import { uploadImageOnCloudinary } from "../../utils/cloudinary.js";
+import { uploadImageOnCloudinary } from "../../utils/shared/cloudinary.js";
 import { cookieOptions } from "../../constants/cookieOptions.js";
-import { checkRequiredFields } from "../../utils/requiredFieldsCheck.js";
+import { checkRequiredFields } from "../../utils/v1/requiredFieldsCheck.js";
 
 const getDeliveryPartnerAccount = asyncHandler(async (req, res) => {
   if (!req.deliveryPartner || !req.deliveryPartner.id) {
