+++ src/controllers/v1/healthCheck.controller.js
@@ -1,5 +1,5 @@
-import { asyncHandler } from "../../utils/asyncHandler.js";
-import { ApiResponse } from "../../utils/apiResponse.js";
+import { asyncHandler } from "../../utils/shared/asyncHandler.js";
+import { ApiResponse } from "../../utils/shared/apiResponse.js";
 import { sql } from "../../database/v1/database.js";
 
 const healthCheck = asyncHandler(async (req, res) => {
