+++ src/middlewares/v1/auth.middleware.js
@@ -1,12 +1,12 @@
 import jwt from "jsonwebtoken";
-import { asyncHandler } from "../../utils/asyncHandler.js";
+import { asyncHandler } from "../../utils/shared/asyncHandler.js";
 import { getNonSensitiveCustomerInfoById } from "../../database/v1/queries/customer.query.js";
 import { getNonSensitiveManagementTeamInfoById } from "../../database/v1/queries/managementTeam.query.js";
 import { getNonSensitiveRestaurantOwnerInfoById } from "../../database/v1/queries/restaurantOwner.query.js";
 import { getNonSensitiveBusinessTeamInfoById } from "../../database/v1/queries/businessTeam.query.js";
 import { getNonSensitiveDeliveryPartnerInfoById } from "../../database/v1/queries/deliveryPartner.query.js";
 import { getRestaurantById } from "../../database/v1/queries/restaurant.query.js";
-import ApiError from "../../utils/apiError.js";
+import ApiError from "../../utils/shared/apiError.js";
 import { STATUS } from "../../constants/statusCodes.js";
 
 export const verifyUserJwt = asyncHandler(async (req, res, next) => {
